dags/requirements.txt
apache-airflow[postgres]==2.3.0
psycopg2-binary==2.9.9
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1
pandas==1.3.5
requests==2.31.0
pyarrow==13.0.0
numpy>=1.24.4
sqlalchemy<1.4.16
apache-airflow-providers-google==8.11.0
google-cloud-storage
selenium